Refactor
Changed names and set welcome email function of Cognito on.
